New Launcher Icon
Updated the launcher icon to match material design style.

There are two versions included in this commit.

* ic_launcher.png
* ic_launcher_flat.png

ic_launcher.png
Includes a white border and a subtle drop shadow. Looks good for use as
the icon to help make it look good on various home screens.

ic_launcher_flat.png
has no border and no drop shadow. Looks good for use in the action bar,
on the "Join the story maker community" screen, and any other place we
need to present the logo on its own and may not want a white border.

Notes on the design: Dropped the internal black ring, thickened the
color border a bit and made the center circle pure white instead of off
white. This is my attempt to make it a little more in line with
material design. Should also make it easier for people with both
versions on their devices.