Misc. Icons
Created some general icons that match the icons I used in the new
artwork for the content packs.

ic_lock
The security lock I used in the cover of the security lessons.

ic_mic
Icon used in the audio lessons.

Would be great to update the icon on the audio clip card from the photo
icon to be this audio recorder next to the capture button.

ic_photo
Icon used in the Photography lessons. Already in use on the photo clip
cards.

Ic_video_sm
Icon used in the video lessons cover images.

ic_videocam
The official android material design icon for video, would be good to
updated the video clip card with this icon for the capture button.